feat(inventory): add inventory module and refactor build params
- Run cargo fmt to format the codebase.
- Implement the core inventory logic.
- Group 7+ parameters into a single struct for better readability and maintainability.

+++ src/models/mod.rs
@@ -1,6 +1,7 @@
-use sqlx::FromRow;
-use serde::{Serialize, Deserialize};
 use chrono::{DateTime, Local};
+use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
+use sqlx::types::BigDecimal;
 
 #[derive(FromRow, Debug, Clone, Serialize, Deserialize)]
 pub struct User {
@@ -17,21 +18,20 @@ pub struct User {
     pub updated_at: Option<DateTime<Local>>,
 }
 
-
-
 #[derive(FromRow, Debug, Clone, Serialize, Deserialize)]
 pub struct Order {
     pub order_id: String,
     pub user_id: i32,
-    pub total_amount: f64,
-    pub pay_amount: Option<f64>,
+    pub total_amount: BigDecimal,
+    pub pay_amount: Option<BigDecimal>,
     pub order_status: i8,
     pub consignee_info: Option<serde_json::Value>,
     pub created_at: Option<DateTime<Local>>,
+    #[serde(skip)]
     pub updated_at: Option<DateTime<Local>>,
+    #[serde(skip)]
     pub deleted_at: Option<DateTime<Local>>,
 }
-
 
 #[derive(FromRow, Debug, Clone, Serialize, Deserialize)]
 pub struct OrderItem {
@@ -40,9 +40,11 @@ pub struct OrderItem {
     pub product_id: i32,
     pub product_name: String,
     pub quantity: i32,
-    pub unit_price: f64,
-    pub subtotal: f64,
+    pub unit_price: BigDecimal,
+    pub subtotal: BigDecimal,
     pub spec_info: Option<String>,
+    #[serde(skip)]
     pub created_at: Option<DateTime<Local>>,
+    #[serde(skip)]
     pub is_deleted: i8,
 }
